refactor: execute logic
