+++ playground/src/App.vue
@@ -4,7 +4,6 @@ import Playground from '~/components/playground/Playground.vue'
 
 <template>
   <div h-screen p="4" class="dark:bg-dark-900 dark:text-light-100">
-
     <Playground />
   </div>
 </template>
