+++ playground/src/components/playground/Editor.vue
@@ -11,32 +11,28 @@ const { onChange, setContent, execute } = useMonaco(target, {
   code: props.value,
 })
 
-const eventBus = useEventBus<string>('code-change')
+type Events = { type: 'EXECUTE_CODE', data: string } | { type: 'COMPILE_CODE' }
+
+const eventBus = useEventBus<Events>('code')
 
 watch(() => props.value, () => setContent(props.value))
+
 onChange(content => emit('change', content))
+
 emit('change', props.value)
 
-function onClick() {
-  execute((code) => {
-    eventBus.emit(code)
-  })
-}
+eventBus.on((e) => {
+  if (e.type === 'COMPILE_CODE') {
+    execute((code) => {
+      eventBus.emit({ type: 'EXECUTE_CODE', data: code })
+    })
+  }
+})
+
 </script>
 
 <template>
   <Pane title="Code">
-    <template #controls>
-      <button
-        type="button"
-        h-8 px-2
-        border="l-1 dark:dark-300"
-        bg="hover:dark-200 active:dark-300"
-        @click="onClick"
-      >
-        Execute
-      </button>
-    </template>
     <div class="h-full w-full of-hidden">
       <div ref="target" class="h-full w-full" />
     </div>
